refactor: use sonner toast directly
