+++ components/ui/index.ts
@@ -46,4 +46,3 @@ export * from "./toggle"
 export * from "./toggle-group"
 export * from "./tooltip"
 export * from "./use-mobile"
-export * from "./use-toast"
